+++ src/components/SessionPanel/__tests__/TimelineButtons.test.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { mount } from 'enzyme';
-import { TimelineButtons } from '../TimelineButtons';
+import { SessionNavigation } from '../SessionNavigation';
 
 
 describe('Timeline component', () => {
@@ -11,7 +11,7 @@ describe('Timeline component', () => {
   const nextMock = jest.fn();
 
   const component = mount(
-    <TimelineButtons
+    <SessionNavigation
       percentProgress="40"
       onClickBack={backMock}
       onClickNext={nextMock}
