+++ src/components/SettingsMenu/Sections/DeviceSettings.js
