+++ src/components/SettingsMenu/Sections/VisualPreferences.js
