--- src/containers/Timeline/index.js
@@ -1,3 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export { default as StagesMenu } from './StagesMenu';
-export { default as SubMenu } from './SubMenu';
